Problem: Users cannot log 'FYI' type problems (#195)
Solution: On the `new problem` page, add a checkbox which marks the problem as an FYI. If a problem is an FYI, don't allow it to be claimed. Instead, show a button "Got it", which appends the user's name to a list of people who have read and understood the problem.