Problem: Problems not marked as stale (#328)
Solution: Add a cron job every day that'll check for stale problems and
mark them accordingly. A problem is considered stale if it hadn't had any
activity in the last 10 days. If the problem is marked as stale, notify
the author. Also, implement logic that will return the problem to its old
status if it gets some activity after being marked as stale.